Bugfix, get_frame gave error when game started
AcquireNextFrame gave DXGI_ERROR_INVALID_CALL when specific game was
started, Mount and Blade Warband. This error means that the previous
frame was not release, but such was not the case. Probably, it had
something to do with the mode change at game startup, and so
refresh_output was added to handle error